Refactor: rename file to chartService.ts
